feat: finished all functionality
